added local storage and changed fetch to user.js
