+++ frontend/src/components/Login.js
@@ -3,6 +3,7 @@ import { useDispatch, useSelector, batch } from 'react-redux'
 import { useHistory, Link } from 'react-router-dom'
 
 import user from '../reducers/user'
+import { sign } from '../reducers/user'
 
 import { API_URL } from '../reusable/urls'
 
@@ -24,28 +25,7 @@ const Login = () => {
 
     const onFormSubmit = (e) => {
         e.preventDefault()
-
-        const options = {
-            method: 'POST',
-            headers:{
-                'Content-type': 'application/json'
-            },
-            body: JSON.stringify({ username, password})
-        }
-        fetch(API_URL(mode), options)
-          .then(res => res.json())
-          .then(data => {
-             if (data.success) {
-                batch(() => {//batch updates all the dispatches at the same time, not one at a time
-                    dispatch(user.actions.setUsername(data.username))
-                    dispatch(user.actions.setAccessToken(data.accessToken))
-                    dispatch(user.actions.setErrors(null))
-                })
-            } else {
-                dispatch(user.actions.setErrors(data))
-            }
-        })
-        .catch()
+        dispatch(sign(username, password, mode))
     }
 
     return (
