+++ frontend/src/reducers/user.js
@@ -1,4 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit'
+import { batch } from 'react-redux'
+import { API_URL } from 'reusable/urls'
 
 const initialState = localStorage.getItem('user')
   ? {
@@ -14,11 +16,7 @@ const initialState = localStorage.getItem('user')
 
 const user = createSlice ({
     name: 'user',
-    initialState: {
-        username: null,
-        accessToken: null,
-        errors: null
-    },
+    initialState,
     reducers: {
         setUsername: (store, action) => {
             store.username = action.payload
@@ -31,5 +29,36 @@ const user = createSlice ({
         }
     } 
 })
+
+export const sign = (username, password, mode) => {
+  return (dispatch, getStore) => {
+    const options = {
+      method:'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({ username, password })
+    }
+
+    fetch(API_URL(mode), options)
+      .then(res => res.json())
+      .then(data => {
+        if (data.success) {
+          batch(() => {
+            dispatch(user.actions.setUsername(data.username))
+            dispatch(user.actions.setAccessToken(data.accessToken))
+            dispatch(user.actions.setErrors(null))
+            
+            localStorage.setItem('user', JSON.stringify({
+                username: data.username,
+                accessToken: data.accessToken
+            }))
+          })
+        } else {
+            dispatch(user.actions.setErrors(data))
+        }
+      })
+  }
+}
 
 export default user
